wip: add tabs for signal
