+++ src/S7SvrSim/App.xaml
@@ -14,25 +14,6 @@
                 <ResourceDictionary Source="pack://application:,,,/MaterialDesignColors;component/Themes/Recommended/Primary/MaterialDesignColor.DeepPurple.xaml" />
                 <ResourceDictionary Source="pack://application:,,,/MaterialDesignColors;component/Themes/Recommended/Accent/MaterialDesignColor.Lime.xaml" />
                 <ResourceDictionary Source="/UserControls/Signals/SignalValueSetBoxTheme.xaml" />
-
-                <ResourceDictionary>
-                    <Style TargetType="Button"
-                           x:Key="TextButton"
-                           >
-                        <Setter Property="Cursor" Value="Hand" />
-                        <Setter Property="Template">
-                            <Setter.Value>
-                                <ControlTemplate>
-                                    <Border Margin="{TemplateBinding Margin}"
-                                            Padding="{TemplateBinding Padding}"
-                                            >
-                                        <ContentPresenter />
-                                    </Border>
-                                </ControlTemplate>
-                            </Setter.Value>
-                        </Setter>
-                    </Style>
-                </ResourceDictionary>
             </ResourceDictionary.MergedDictionaries>
         </ResourceDictionary>
