+++ src/S7SvrSim/S7Signal/SignalTypeCache.cs
@@ -13,5 +13,10 @@ namespace S7SvrSim.Services
         }
 
         public Type[] Value { get; }
+
+        public void Write(Type[] value)
+        {
+
+        }
     }
 }
