+++ src/S7SvrSim/Services/IMemCache.cs
@@ -3,5 +3,7 @@
     public interface IMemCache<T>
     {
         T Value { get; }
+
+        void Write(T value);
     }
 }
